// web/packages/design/src/tokens/generated/fluid.stylex.ts
// Generated by scripts/generate.ts. Do not edit; run `pnpm generate`.
import * as stylex from "@stylexjs/stylex";

export const fluidType = stylex.defineVars({
  stepn2: "clamp(0.6944rem, 0.6859rem + 0.0426vw, 0.72rem)",
  stepn1: "clamp(0.8333rem, 0.8111rem + 0.1111vw, 0.9rem)",
  step0: "clamp(1rem, 0.9583rem + 0.2083vw, 1.125rem)",
  step1: "clamp(1.2rem, 1.1312rem + 0.3438vw, 1.4063rem)",
  step2: "clamp(1.44rem, 1.3341rem + 0.5297vw, 1.7578rem)",
  step3: "clamp(1.728rem, 1.5716rem + 0.7821vw, 2.1973rem)",
  step4: "clamp(2.0736rem, 1.8493rem + 1.1216vw, 2.7466rem)",
  step5: "clamp(2.4883rem, 2.1734rem + 1.5748vw, 3.4332rem)",
  step6: "clamp(2.986rem, 2.5508rem + 2.1759vw, 4.2915rem)",
});

export const fluidSpace = stylex.defineVars({
  xs: "clamp(0.5rem, 0.4583rem + 0.2083vw, 0.625rem)",
  sm: "clamp(0.75rem, 0.6667rem + 0.4167vw, 1rem)",
  md: "clamp(1rem, 0.8333rem + 0.8333vw, 1.5rem)",
  lg: "clamp(1.5rem, 1.25rem + 1.25vw, 2.25rem)",
  xl: "clamp(2rem, 1.5rem + 2.5vw, 3.5rem)",
  xxl: "clamp(3rem, 2.1667rem + 4.1667vw, 5.5rem)",
  xxxl: "clamp(4rem, 2.6667rem + 6.6667vw, 8rem)",
});